% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{collect_region}
\alias{collect_region}
\title{Collecting Region Ingroup Share}
\usage{
collect_region(group, region, rs)
}
\arguments{
\item{group}{Integer. Group membership of the agent.}

\item{region}{Integer. Region the agent resides.}

\item{rs}{\code{region_stat} exported in the list generated by \code{\link{initialsoc}} function.}
}
\value{
Numeric. The share of ingroup members in the region.
}
\description{
Recoriding the share of ingroup members in region for each agent.
}
