% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{collect_subregion}
\alias{collect_subregion}
\title{Collecting Sub-Region Ingroup Share}
\usage{
collect_subregion(group, subregion, rs)
}
\arguments{
\item{group}{Integer. Group membership of the agent.}

\item{subregion}{Integer. Sub-region the agent resides.}

\item{rs}{\code{subregion_stat} exported in the list generated by \code{\link{initialsoc}} function.}
}
\value{
Numeric. The share of ingroup members in the region.
}
\description{
Recoriding the share of ingroup members in sub-region for each agent.
}
