% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Collecting Neighbor Locations}
\usage{
get_neighbors(coords, radius, spacelim, includeme = FALSE)
}
\arguments{
\item{coords}{Coordinates of the agent.}

\item{radius}{Search radius of neighbor.}

\item{spacelim}{Integer vector of length 2. Grid space size 
(number of rows and number of columns, respectively).}

\item{includeme}{If TRUE, collect the agent's own value.}
}
\value{
Coordinates of neightbors
}
\description{
Collecting the location of neighbors.
}
