% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{getcontext}
\alias{getcontext}
\title{Get Neighborhood Context}
\usage{
getcontext(soc, personloc, radius, timetoagg = 2)
}
\arguments{
\item{soc}{The lists generated initially by \code{\link{initialsoc}}. 
Values of previous elections votes (i.e., \code{"grid_votes1"} and \code{"grid_votes2"}) must be included.}

\item{personloc}{The coordinates with agents.}

\item{radius}{Search radius to collect neighborhood preferences.}

\item{timetoagg}{How many previous elections to average. Either 1 or 2 (default).}
}
\value{
The list with:
\itemize{
  \item \code{local} Local vote share in previous elections for voters 
  (i.e., the voting patterns of neightbors in search radius)
  \item \code{national} National vote share in previous elections. 
  \item \code{detail} Detailed results.
}
}
\description{
Get neighborhood context from society data
}
