% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{getregion}
\alias{getregion}
\title{Get Regional Context for Voting}
\usage{
getregion(soc, personloc, timetoagg = 2)
}
\arguments{
\item{soc}{The lists generated initially by \code{\link{initialsoc}}. 
Values of previous elections votes (i.e., \code{"grid_votes1"} and \code{"grid_votes2"}) must be included.}

\item{personloc}{The coordinates with agents.}

\item{timetoagg}{How many previous elections to average. Either 1 or 2 (default).}
}
\value{
The list with:
\itemize{
  \item \code{sublocal} Sub-regional vote share in previous elections (Standard contextual voting Used).
  \item \code{local} Regional vote share in previous elections (Standard contextual voting used).
  \item \code{national} National vote share in previous elections (Standard contextual voting used). 
  \item \code{sublocal1} Sub-regional vote share in previous elections (Local only contextual voting Used).
  \item \code{local1} Regional vote share in previous elections (Local only contextual voting used).
  \item \code{national1} National vote share in previous elections (Local only contextual voting used). 
  \item \code{sublocal2} Sub-regional vote share in previous elections (National only contextual voting Used).
  \item \code{local2} Regional vote share in previous elections (National only contextual voting used).
  \item \code{national2} National vote share in previous elections (National only contextual voting used). 
}
}
\description{
Get regional context from society data
}
