% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{getsignal}
\alias{getsignal}
\title{Get Preference Signals}
\usage{
getsignal(party, ideology, knowledge, difficulty = 3, bias_common = 0,
  bias_ingroup = 0, pastbias = NULL, idedist_blue_signal_seed = NULL,
  idedist_red_signal_seed = NULL, capacity_signal_seed = NULL)
}
\arguments{
\item{party}{The lists generated initially by \code{\link{setparty}} and may be updated by \code{\link{updateparty}}.}

\item{ideology}{The lists generated initially by \code{\link{setideology}} and may be updated by \code{\link{updateideology}}.}

\item{knowledge}{The lists generated by \code{\link{setknowledge}}.}

\item{difficulty}{The extent of noise level in signal.}

\item{bias_common}{The extent of common directional bias (i.e., positive value advantages red party and vice versa) in signal.}

\item{bias_ingroup}{The extent of common ingroup bias (i.e., Evaluate the ingroup candidate better).}

\item{pastbias}{Currently not used.}

\item{idedist_blue_signal_seed}{The random number seed to draw the signal regarding ideological distance from blue party.}

\item{idedist_red_signal_seed}{The random number seed to draw the signal regarding ideological distance from red party.}

\item{capacity_signal_seed}{The random number seed to draw the signal regarding the capacity of incumbent party.}
}
\value{
The list of signal parameters: 
\itemize{
  \item \code{signal} The list of signals regarding ideological distance from parties and incumbent capacity.
  \item \code{truth} The list of truth regarding ideological distance from parties and incumbent capacity.
  \item \code{seed} Random number seeds used.
}
}
\description{
Obtain preference signals for voters.
}
\examples{

# initiate ideology, knowledge and party
set.seed(30)
x <- sample(rep(c(-1,1),each=1000))
ide <- setideology(x, 0, 0.8, 1)
kn <- setknowledge(x, 0.60, 0.70, 0.85)
par <- setparty()

# Get signals
sig <- getsignal(par,ide,kn,3) # Difficulty of 3.3 creates approx. 1.9 SD
sd(sig$signal$idedist_blue) # around 1.9
sd(sig$signal$idedist_red) # around 1.9
par(mfrow=c(2,2),mar=c(2,2,2,2))
plot(sig$truth$idedist_blue,sig$signal$idedist_blue)
plot(sig$truth$idedist_red,sig$signal$idedist_red)
plot(sig$truth$idedist_red[kn$p>0.6],
     sig$signal$idedist_red[kn$p>0.6])
plot(sig$truth$idedist_red[kn$p<0.6],
     sig$signal$idedist_red[kn$p<0.6])

}
