% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{happyregion}
\alias{happyregion}
\title{Stage Assessment to Aggregate Happiness by Region}
\usage{
happyregion(soc)
}
\arguments{
\item{soc}{The list of society characteristics initiated by \code{\link{initialsoc}}.}
}
\value{
The list of society characteristics with the regional in-group share measures.
Update \code{region_stat} and generate \code{ingroupshare} and \code{grid_ingroupshare}.
}
\description{
Aggregating in-group share status of agents by region
}
