% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{happysoc}
\alias{happysoc}
\title{Stage Assessment to Aggregate Happiness by Neighborhood}
\usage{
happysoc(soc, radius, alike_threshold, initiate = FALSE)
}
\arguments{
\item{soc}{The list of society characteristics initiated by \code{\link{initialsoc}}.}

\item{radius}{Integer. Search radius to look for neighbors.}

\item{alike_threshold}{1 - tolerance in the segregation model.}

\item{initiate}{Default is \code{FALSE}. If \code{TRUE}, do not produce \code{happy_tracker}.}
}
\value{
The list of society characteristics with the happiness measures. 
\code{happy_indicators} records the proportion of happy agents and
\code{happy_tracker} indicates the history of \code{happy_indicators}.
}
\description{
Aggregating the happiness status of agents in the society.
}
