% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{happysubregion}
\alias{happysubregion}
\title{Stage Assessment to Aggregate Happiness by Sub-Region}
\usage{
happysubregion(soc)
}
\arguments{
\item{soc}{The list of society characteristics initiated by \code{\link{initialsoc}}}
}
\value{
The list of society characteristics with the regional in-group share measures.
Update \code{subregion_stat} and generate \code{ingroupshare_sub} and \code{grid_ingroupshare_sub}.
}
\description{
Aggregating in-group share status of agents by sub-region
}
