% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{initialsoc}
\alias{initialsoc}
\title{Initialize the society}
\usage{
initialsoc(number = 2300, sizegroup = c(0.5, 0.5), spacelim = c(54,
  54), regionlevel = c(3, 3), subregionlevel = c(9, 9),
  seedval = 356)
}
\arguments{
\item{number}{Interger. Number of agents.}

\item{sizegroup}{Numeric Vector. Proportion of agents in each group. 
Length defines the number of groups. Must adds up to 1.}

\item{spacelim}{Integer vector of length 2. Grid space size 
(number of rows and number of columns, respectively).}

\item{regionlevel}{Integer vector of length 2. Level to Split Region (Horizontaly & Vertically).}

\item{subregionlevel}{Ingeger vector of length 2. Level to Split Sub-Region (Horizontaly & Vertically).}

\item{seedval}{Random seed value for the simulation.}
}
\value{
List object with society characteristics.
\itemize{
  \item \code{grid} The grid of voters (with group membership at each cell)
  \item \code{grid_region} The grid of region identifier at each cell.
  \item \code{grid_subregion} The grid of sub-region identifier at each cell. 
  \item \code{grid_coords} All grid coordinates. 
  \item \code{group} The numeric vector of group memberships 
  \item \code{region_stat} Aggregated group statistics by region. 
  \item \code{subregion_stat} Aggregated group statistics by region.
}
}
\description{
Initializing the Society for Segregation Model
}
