% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{letsvote}
\alias{letsvote}
\title{Vote in Election}
\usage{
letsvote(soc, b_idedist = 0.198, b_capacity = 0.716,
  b_incumbent = 1.526, b_pvi_nation = -9.827, b_pvi_local = 5.156,
  b_pvi_sublocal = 1.663, b_pid = 1.215, vote_seed = NULL)
}
\arguments{
\item{soc}{The lists generated initially by \code{\link{initialsoc}}. 
Also following elements must be contained within \code{soc} object:
\itemize{
 \item \code{knowledge} Created by \code{\link{setknowledge}}
 \item \code{ideology} Created by \code{\link{setideology}}, may be updated by \code{\link{updateideology}}
 \item \code{context} Created by \code{\link{getregion}}
 \item \code{signal} Created by \code{\link{getsignal}}
 \item \code{incumbent} Character of incumbent party. Either \code{"blue"} or \code{"red"}.
}}

\item{b_idedist}{The influence of ideological distance. The default = 0.198, # From ANES}

\item{b_capacity}{The influence of capacity The default = 0.716, # From ANES}

\item{b_incumbent}{The influence if incumebent. The default = 1.526, # From ANES (EDV result)}

\item{b_pvi_nation}{The influence of national PVI. The default = -9.827, # From ANES (EDV result)}

\item{b_pvi_local}{The influence of regional PVI. The default = 5.156, # From CCES pvi_state}

\item{b_pvi_sublocal}{The influence of sub-regional PVI. The default = 1.663, # From CCES pvi_state}

\item{b_pid}{The influence of group membership. The default = 1.215, # From ANES}

\item{vote_seed}{The random number seed.}
}
\value{
The list with:
\itemize{
  \item \code{vote} The list of vote decisions by different voting functions.
  \item \code{prob} The list of vote probabilities by different voting functions.
  \item \code{pref} The list of raw vote preferences by different voting functions. 
  \item \code{noidprob} The list of vote probabilities (if group membership has no effect).
  \item \code{b} The list of parameters.
  \item \code{tracker} Tracker of voting history.
  \item \code{vote_seed} The random number seed used. 
}
}
\description{
Voters make voting decisions.
}
