% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{moveunhappy}
\alias{moveunhappy}
\title{Move unhappy agents}
\usage{
moveunhappy(soc, seedval = NULL)
}
\arguments{
\item{soc}{The list of society characteristics initiated by \code{\link{initialsoc}}.}

\item{seedval}{The seeding value to control randomization.}
}
\value{
The list of society characteristics with the updated locations of agents.
}
\description{
Move unhappy agents from the current cell.
}
