% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{setideology}
\alias{setideology}
\title{Set Continuous Ideology}
\usage{
setideology(group, commonfactor = 0, groupfactor = 1,
  personalfactor = 1, groupfactor_type = "constant", b_seed = NULL,
  e_seed = NULL)
}
\arguments{
\item{group}{The numeric vector of the group memberships of the agents.}

\item{commonfactor}{The common factor in ideology (intercept).}

\item{groupfactor}{The group influence on ideology.}

\item{personalfactor}{The random personal factor in  ideology (The SD of the random error to the model).}

\item{groupfactor_type}{If \code{"constant"}, \code{groupfactor} is the constant coefficient for \code{group}.
If \code{"diverging"}, \code{groupfactor} is the SD for folded normal distribution that randomly 
assigns coeffcient for \code{group}.}

\item{b_seed}{The random number seed to draw group influence used when \code{groupfactor_type=="constant"}.}

\item{e_seed}{The random number seed to draw personal influence.}
}
\value{
The list of following elements: 
\itemize{
  \item \code{y} The agents' ideologies/
  \item \code{x} The group values rescaled in -1 to 1 range.
  \item \code{a} The common influence = \code{commmonfactor}
  \item \code{b} The vector of the group influcence
  \item \code{e} The vector of the personal influence
  \item \code{tracker} The tracker table of ideology.
}
}
\description{
Initialize ideology of agents with the linear model.
}
\details{
The ideology for agent $i$ $(y_i)$ is calculated by the linear function: 
$y = a + b_i * x_i + e_i$.
}
\examples{

set.seed(30)
x <- sample(rep(c(-1,1),each=1000))
table(x)

# Ideology Highly Sorted by Group
ide <- setideology(x, 0, 1.5, 1)
par(mfrow=c(1,3),mar=c(2,2,2,2))
hist(ide$y, xlim=c(-6,6)); mean(ide$y); sd(ide$y)
hist(ide$y[x==-1],xlim=c(-6,6)); sd(ide$y[x==-1])
hist(ide$y[x==1],xlim=c(-6,6)); sd(ide$y[x==1])

# Ideology Moderately Sorted by Group
ide <- setideology(x, 0, 1, 1)
par(mfrow=c(1,3),mar=c(2,2,2,2))
hist(ide$y, xlim=c(-6,6)); mean(ide$y); sd(ide$y)
hist(ide$y[x==-1],xlim=c(-6,6)); sd(ide$y[x==-1])
hist(ide$y[x==1],xlim=c(-6,6)); sd(ide$y[x==1])

# Ideology Weakly Sorted by Group
ide <- setideology(x, 0, 0.8, 1)
par(mfrow=c(1,3),mar=c(2,2,2,2))
hist(ide$y, xlim=c(-6,6)); mean(ide$y); sd(ide$y)
hist(ide$y[x==-1],xlim=c(-6,6)); sd(ide$y[x==-1])
hist(ide$y[x==1],xlim=c(-6,6)); sd(ide$y[x==1])

}
