% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{setknowledge}
\alias{setknowledge}
\title{Set Knowledge}
\usage{
setknowledge(group, mean1 = 0.6, mean2 = 0.7, spread = 0.8,
  kn_seed = NULL)
}
\arguments{
\item{group}{The vector of group membership of agents.}

\item{mean1}{The lowest average level of knowledge for a group.}

\item{mean2}{The highest average level of knowledge for a group.}

\item{spread}{The spread (SD for normal distribution) of knowledg within a group.}

\item{kn_seed}{Random seed value to draw knowledge}
}
\value{
The updated list of ideology parameters 
\itemize{
  \item \code{knp} Knowledge level converted to 0-1 range by inverse-standard-normal function. 
  \item \code{kn} The raw knowledge score
  \item \code{kn_seed} random seed value used.
}
}
\description{
Set knowledge levels of agents (Now only applicable with two groups).
}
\details{
If two groups, the raw knowledge score is generated by the normal distribution with 
mean \code{mean1} and standard deviation \code{spread} for the first group; 
the normal distribution with mean \code{mean1} and standard deviation \code{spread} 
for the second group. If more than two groups, additional group means are assigned by the 
points that equally divides the interval between \code{mean1} and \code{mean2}.
}
\examples{

# initiate groups
set.seed(30)
x <- sample(rep(c(-1,1),each=1000))

# Set Knowledge

# sd of 0.85 produce approx. 0.25 SD after conversion.
kn <- setknowledge(x, 0.65, 0.65, 0.85)
par(mfrow=c(2,2),mar=c(2,2,2,2))
hist(kn$p); mean(kn$p); sd(kn$p)
hist(kn$y); mean(kn$y); sd(kn$y)
hist(kn$p[x==-1]); mean(kn$p[x==-1]); sd(kn$p[x==-1])
hist(kn$p[x==1]); mean(kn$p[x==1]); sd(kn$p[x==1])

# sd of 0.8 produce approx. 0.25 SD after conversion.
kn <- setknowledge(x, 0.60, 0.70, 0.85)
par(mfrow=c(2,2),mar=c(2,2,2,2))
hist(kn$p); mean(kn$p); sd(kn$p)
hist(kn$y); mean(kn$y); sd(kn$y)
hist(kn$p[x==-1]); mean(kn$p[x==-1]); sd(kn$p[x==-1])
hist(kn$p[x==1]); mean(kn$p[x==1]); sd(kn$p[x==1])

}
