% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{setparty}
\alias{setparty}
\title{Set Characteristics of Two Parties}
\usage{
setparty(blue_startloc = -0.8, red_startloc = 0.8,
  startloc_noise = 0.1, blue_startcapacity = 0,
  red_startcapacity = 0, startcapacity_noise = 0.1,
  startincumbent = NULL, blue_loc_seed = NULL, red_loc_seed = NULL,
  blue_capacity_seed = NULL, red_capacity_seed = NULL)
}
\arguments{
\item{blue_startloc}{Starting location of blue party ideology.}

\item{red_startloc}{Starting location of red party ideology.}

\item{startloc_noise}{Random noise given to the starting locations of party ideologies.}

\item{blue_startcapacity}{Starting level of blue party capacity.}

\item{red_startcapacity}{Starting level of red party capacity.}

\item{startcapacity_noise}{Random noise given to the starting level of party capacities.}

\item{startincumbent}{Incumbent in the first election. If \code{NULL}, randomly chosen.}

\item{blue_loc_seed}{Random number seed to draw blue party ideology.}

\item{red_loc_seed}{Random number seed to draw red party ideology.}

\item{blue_capacity_seed}{Random number seed to draw blue party capacity.}

\item{red_capacity_seed}{Random number seed to draw red party capacity.}
}
\value{
The updated list of ideology parameters 
\itemize{
  \item \code{loc} ideology locations of two parties. 
  \item \code{capacity} capacity levels of two parties.
  \item \code{tracker} Tracker of the history of ideogoloy and capacity.
}
}
\description{
Set ideologies and capacity of parties (i.e., blue and red) running in election.
}
