% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{updateideology}
\alias{updateideology}
\title{Update Continuous Ideology}
\usage{
updateideology(ideology, commonfactor = 0, groupfactor = 0,
  personalfactor = 0.2, b_seed = NULL, e_seed = NULL)
}
\arguments{
\item{ideology}{The list of ideology parameters initiated by \code{\link{setideology}}.}

\item{commonfactor}{The movement in common factor.}

\item{groupfactor}{The movement in group influence on ideology.}

\item{personalfactor}{The SD for the movemebnt in personal influence on ideology.}

\item{b_seed}{The random number seed to draw group influence used when 
\code{groupfactor_type=="constant"} when the ideolog is initiated.}

\item{e_seed}{The random number seed to draw personal influence.}
}
\value{
The updated list of ideology parameters 
\itemize{
  \item \code{y} The agents' ideologies. 
  \item \code{x} The group values rescaled in -1 to 1 range.
  \item \code{a} $a_{t-1} + $\code{commonfactor}
  \item \code{b} $b_{t-1} + $\code{groupfactor} or $b_{i,t-1} + abs(N(0,abs(groupfactor)))$
  \item \code{e} $e_{i, t-1} + N(0, personalfactor)$ 
  \item \code{tracker} The tracker table of ideology.
}
}
\description{
Update ideology of agents with the linear model.
}
\details{
The ideology for agent $i$ $(y_i)$ is recalculated by the linear function: 
$y = a + b_i * x_i + e_i$.
}
\examples{

# initiate ideology
set.seed(30)
x <- sample(rep(c(-1,1),each=1000))
ide <- setideology(x, 0, 0.8, 1)

# update
idex <- updateideology(ide, 0)
par(mfrow=c(1,3),mar=c(2,2,2,2))
hist(ide$y); mean(ide$y); sd(ide$y)
hist(idex$y); mean(idex$y); sd(idex$y)
plot(ide$y, idex$y);

}
