% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_functions_latest.R
\name{updateparty}
\alias{updateparty}
\title{Update Characteristics of Two Parties}
\usage{
updateparty(party, incumbent, loc_noise = 0, capacity_noise = 0.7,
  blue_loc_direction = 0, red_loc_direction = 0,
  blue_capacity_direction = 0, red_capacity_direction = 0,
  update_type = "oneshot", blue_loc_seed = NULL, red_loc_seed = NULL,
  blue_capacity_seed = NULL, red_capacity_seed = NULL)
}
\arguments{
\item{party}{The lists generated initially by \code{\link{setparty}}.}

\item{incumbent}{Incumbent party in the current election. Either \code{"blue"} or \code{"red"}.}

\item{loc_noise}{Random noise in updating locations of party ideologies.}

\item{capacity_noise}{Random noise in updating party capacity.}

\item{blue_loc_direction}{Direction and expected quantity of the movement in blue party ideology.}

\item{red_loc_direction}{Direction and expected quantity of the movement in red party ideology.}

\item{blue_capacity_direction}{Direction and expected quantity of the movement in blue party capacity.}

\item{red_capacity_direction}{Direction and expected quantity of the movement in red party capacity.}

\item{update_type}{The type of updating procedure. 
If \code{"oneshot"}, the update is always given over \code{startloc} and \code{startcapacity} defined in 
\code{\link{setparty}} at each election. If \code{"online"}, the update is given over the ideology and 
capacity realized in the previous election. If \code{"fromstart"} the update is given over the 
ideology and capacity realized in the first election of the simulation run.}

\item{blue_loc_seed}{Random number seed to draw blue party ideology.}

\item{red_loc_seed}{Random number seed to draw red party ideology.}

\item{blue_capacity_seed}{Random number seed to draw blue party capacity.}

\item{red_capacity_seed}{Random number seed to draw red party capacity.}
}
\value{
The updated list of ideology parameters: 
\itemize{
  \item \code{loc} ideology locations of two parties. 
  \item \code{capacity} capacity levels of two parties.
  \item \code{tracker} Tracker of the history of ideogoloy and capacity.
}
}
\description{
Set ideologies and capacity of parties (i.e., blue and red) running in election.
}
\examples{

require(ggplot2)

par <- setparty()
for (t in 1:14) {
  incumbent <- sample(c("red","blue"),1)
  par <- updateparty(par,incumbent, 0, 0.7)
}
# party
ggplot(par$tracker,aes(x=t)) +
  geom_line(aes(y=blue_loc,color="2.blue")) +
  geom_line(aes(y=red_loc,color="1.red")) +
  theme_bw()
# capacity
ggplot(par$tracker,aes(x=t)) +
  geom_line(aes(y=blue_capacity,color="2.blue")) +
  geom_line(aes(y=red_capacity,color="1.red")) +
  theme_bw()

}
